requirements.txt
quantum-grove
pyquil